requirements.txt
Pyrofork
TgCrypto
Pyleaves
python-dotenv
psutil